+++ src/providers/extractors/movieextractor.js
@@ -30,7 +30,7 @@ export default class Extractor extends BaseExtractor {
      * @type {Util}
      */
     this._util = new Util();
-  };
+  }
 
   /**
    * Get all the movies.
@@ -44,7 +44,7 @@ export default class Extractor extends BaseExtractor {
     } catch (err) {
       return this._util.onError(err);
     }
-  };
+  }
 
   /**
    * Extract movie information based on a regex.
@@ -87,7 +87,7 @@ export default class Extractor extends BaseExtractor {
     };
 
     return movie;
-  };
+  }
 
   /**
    * Get movie info from a given torrent.
@@ -108,7 +108,7 @@ export default class Extractor extends BaseExtractor {
     } else {
       logger.warn(`${this.name}: Could not find data from torrent: '${torrent.title}'`);
     }
-  };
+  }
 
   /**
    * Puts all the found movies from the torrents in an array.
@@ -147,7 +147,7 @@ export default class Extractor extends BaseExtractor {
     } catch (err) {
       return this._util.onError(err);
     }
-  };
+  }
 
   /**
    * Returns a list of all the inserted torrents.
@@ -169,6 +169,6 @@ export default class Extractor extends BaseExtractor {
     } catch (err) {
       this._util.onError(err);
     }
-  };
+  }
 
-};
+}