+++ src/providers/helpers/moviehelper.js
@@ -3,7 +3,7 @@ import asyncq from "async-q";
 
 import Movie from "../../models/Movie";
 import Util from "../../util";
-import { fanart, omdb, trakt } from "../../config/constants";
+import { fanart, omdb, tmdb, trakt } from "../../config/constants";
 
 /** Class for saving movies. */
 export default class Helper {
@@ -24,7 +24,7 @@ export default class Helper {
      * @type {Util}
      */
     this._util = new Util();
-  };
+  }
 
   /**
    * Update the torrents for an existing movie.
@@ -58,7 +58,7 @@ export default class Helper {
 
     if (update) movie.torrents[language][quality] = found.torrents[language][quality];
     return movie;
-  };
+  }
 
   /**
    * @description Update a given movie.
@@ -92,7 +92,7 @@ export default class Helper {
     } catch (err) {
       return this._util.onError(err);
     }
-  };
+  }
 
   /**
    * Adds torrents to a movie.
@@ -103,12 +103,12 @@ export default class Helper {
   addTorrents(movie, torrents) {
     return asyncq.each(Object.keys(torrents), torrent => movie.torrents[torrent] = torrents[torrent])
       .then(() => this._updateMovie(movie));
-  };
+  }
 
   /**
    * Get images from themoviedb.org or omdbapi.com.
    * @param {Integer} tmdb_id - The tmdb id of the movie you want the images from.
-   * @param {Integer} imdb_id - The imdb id of the movie you want the images from.
+   * @param {String} imdb_id - The imdb id of the movie you want the images from.
    * @returns {Object} - Object with a banner, fanart and poster images.
    */
   async _getImages(tmdb_id, imdb_id) {
@@ -120,10 +120,17 @@ export default class Helper {
     };
 
     try {
-      const fanartImages = await fanart.getMovieImages(tmdb_id);
-      images.banner = fanartImages.moviebanner ? fanartImages.moviebanner[0].url : holder;
-      images.fanart = fanartImages.moviebackground ? fanartImages.moviebackground[0].url : fanartImages.hdmovieclearart ? fanartImages.hdmovieclearart[0].url : holder;
-      images.poster = fanartImages.movieposter ? fanartImages.movieposter[0].url : holder;
+      const tmdbData = await tmdb.call(`/movie/${tmdb_id}/images`, {});
+
+      let tmdbPoster = tmdbData['posters'].filter(poster => poster.iso_639_1 === "en" || poster.iso_639_1 === null)[0]
+      tmdbPoster = tmdb.getImageUrl(tmdbPoster.file_path, 'w500');
+
+      let tmdbBackdrop = tmdbData['backdrops'].filter(backdrop => backdrop.iso_639_1 === "en" || backdrop.iso_639_1 === null)[0];
+      tmdbBackdrop = tmdb.getImageUrl(tmdbBackdrop.file_path, 'w500');
+
+      images.banner = tmdbPoster ? tmdbPoster : holder;
+      images.fanart = tmdbBackdrop ? tmdbBackdrop : holder;
+      images.poster = tmdbPoster ? tmdbPoster : holder;
     } catch (err) {
       try {
         const omdbImages = await omdb.byID({
@@ -134,7 +141,14 @@ export default class Helper {
         images.fanart = omdbImages.Poster? omdbImages.Poster : holder;
         images.poster = omdbImages.Poster ? omdbImages.Poster : holder;
       } catch (err) {
-        return this._util.onError(`Images: Could not find images on: ${err.path || err} with id: '${tmdb_id || imdb_id}'`);
+        try {
+          const fanartImages = await fanart.getMovieImages(tmdb_id);
+          images.banner = fanartImages.moviebanner ? fanartImages.moviebanner[0].url : holder;
+          images.fanart = fanartImages.moviebackground ? fanartImages.moviebackground[0].url : fanartImages.hdmovieclearart ? fanartImages.hdmovieclearart[0].url : holder;
+          images.poster = fanartImages.movieposter ? fanartImages.movieposter[0].url : holder;
+        } catch (err) {
+          return this._util.onError(`Images: Could not find images on: ${err.path || err} with id: '${tmdb_id || imdb_id}'`);
+        }
       }
     }
 
@@ -178,7 +192,7 @@ export default class Helper {
           images: await this._getImages(traktMovie.ids["tmdb"], traktMovie.ids["imdb"]),
           genres: traktMovie.genres !== null ? traktMovie.genres : ["unknown"],
           released: new Date(traktMovie.released).getTime() / 1000.0,
-          trailer: traktMovie.trailer || false,
+          trailer: traktMovie.trailer || null,
           certification: traktMovie.certification,
           torrents: {}
         };
@@ -186,6 +200,6 @@ export default class Helper {
     } catch (err) {
       return this._util.onError(`Trakt: Could not find any data on: ${err.path || err} with slug: '${slug}'`);
     }
-  };
+  }
 
-};
+}